Move assumeStandardDecimalSeparator into separate object to fix scaladoc
