+++ scala-package/core/src/test/scala/org/apache/mxnet/ModuleSuite.scala
@@ -167,7 +167,7 @@ class ModuleSuite extends FunSuite with BeforeAndAfterAll {
   }
 
   test ("module reshape") {
-    TestUtil.assumeStandardDecimalSeparator()
+    CancelTestUtil.assumeStandardDecimalSeparator()
 
     var sym = Symbol.Variable("data")
     sym = Symbol.FullyConnected("fc")()(Map("data" -> sym, "num_hidden" -> 20))
