+++ scala-package/core/src/test/scala/org/apache/mxnet/TestUtil.scala
@@ -17,12 +17,6 @@
 
 package org.apache.mxnet
 
-import java.text.DecimalFormatSymbols
-import java.util.Locale
-
-import org.scalatest.Assertions
-
-
 class TestUtil {
 
   /**
@@ -50,34 +44,4 @@ class TestUtil {
     getDataDirectory + "/" + relFile
   }
 
-}
-
-object TestUtil {
-
-  /**
-   * Cancel the test if the system's locale uses a decimal separator other than '.'. Please see
-   * #18097 for more information.
-   */
-  def assumeStandardDecimalSeparator(): Unit = {
-    val lcNumeric = System.getenv("LC_NUMERIC");
-
-    val decimalFormatSymbols = if (lcNumeric != null) {
-      val localeName = lcNumeric.stripSuffix(".UTF-8".stripSuffix(".utf-8"))
-      val locale = Locale.forLanguageTag(localeName)
-      DecimalFormatSymbols.getInstance(locale)
-    } else {
-      DecimalFormatSymbols.getInstance()
-    }
-
-    val isStandardDecimalPoint = (decimalFormatSymbols.getDecimalSeparator == '.') &&
-      (lcNumeric != null && lcNumeric.toLowerCase != "en_dk.utf-8") // Java doesn't seem to respect
-                                                                    // the decimal separator
-                                                                    // set in en_DK.UTF8, which is
-                                                                    // used in CentOS CI jobs.
-    if (!isStandardDecimalPoint) {
-      Assertions.cancel("Some operators " +
-        "break when the decimal separator is set to anything other than \".\". These operators " +
-        "should be rewritten to utilize the new FFI. Please see #18097 for more information.")
-    }
-  }
 }
