+++ scala-package/core/src/test/scala/org/apache/mxnet/train/ConvSuite.scala
@@ -32,7 +32,7 @@ class ConvSuite extends FunSuite with BeforeAndAfterAll {
   private var tu = new TestUtil
 
   test("train mnist") {
-    TestUtil.assumeStandardDecimalSeparator()
+    CancelTestUtil.assumeStandardDecimalSeparator()
     // symbol net
     val batchSize = 100
